+++ history.md
@@ -3,6 +3,7 @@
   + open.commons.spring.web.ac: 자원 유형별 권한 정의 어노테이션
   + open.commons.spring.web.aspect: 자원 유형별 권한 중개
   + open.commons.spring.web.beans.factory: 자원 유형 어노테이션 설정 검증.
+  + open.commons.spring.web.autoconfiguration: 자동설정 적용.
 
 [2025/05/14]
 - Patch